% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstats_processing.R
\name{compare.MSstats}
\alias{compare.MSstats}
\title{Perform an MSstats comparison}
\usage{
compare.MSstats(processed.features, comparison.matrix,
  filepath = "MSstats_comparisons.rds", recompute = F)
}
\arguments{
\item{processed.features}{Object with normalized data. Can be produced with process.MSstats()}

\item{comparison.matrix}{A comparison matrix as described in MSstats user guide.}

\item{recompute}{Boolean on whether to recompute comparisons. Default is False}

\item{file.path}{Path to save the resulting comparison data object}
}
\value{
An object containing comparison results
}
\description{
Performs MSstats differential analysis. Saves result to disk and reloads unless recompute is True
}
\examples{
pd.comparisons <- compare.MSstats(rd.proposed, comparisons, recompute=F)

}
