\name{CompleteTriangle}
\alias{CompleteTriangle}
\title{CompleteTriangle}
\usage{
  CompleteTriangle(objProjection)
}
\arguments{
  \item{objProjection}{}
}
\value{
  A data frame
}
\description{
  This function will bind the projected values to the base
  triangle data for a "complete" triangle. This facilitates
  comparison of ultimates between multiple TriangleModels.
}

