\name{CreateCumulative}
\alias{CreateCumulative}
\title{Create cumulative}
\usage{
  CreateCumulative(dfTriangleData, measureCols, Groups)
}
\arguments{
  \item{OriginPeriod}{Vector of intervals}

  \item{Group}{A vector of groups}

  \item{Variable}{A list of}
}
\value{
  Adjusted triangle
}
\description{
  Create cumulative
}
\seealso{
  \code{\link{CreateIncrementals}},
  \code{\link{CreatePriors}}
}

