\name{CreateIncrementals}
\alias{CreateIncrementals}
\title{Create incrementals}
\usage{
  CreateIncrementals(dfTriangleData, measureCols, Groups)
}
\arguments{
  \item{OriginPeriod}{Vector of intervals}

  \item{Group}{A vector of groups}

  \item{Variable}{A list of}

  \item{StartDay}{If OriginStart and OriginEnd are
  supplied, this argument is ignored.}

  \item{StartMonth}{If OriginStart and OriginEnd are
  supplied, this argument is ignored.}
}
\value{
  A vector of intervals
}
\description{
  Create incrementals
}
\seealso{
  \code{\link{CreateCumulative}},
  \code{\link{CreatePriors}}
}

