% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDevelopmentLags.R
\name{CreateLagsFromEvalDates}
\alias{CreateLagsFromEvalDates}
\title{Create Lags from Evaluatino Dates}
\usage{
CreateLagsFromEvalDates(EvalDates, OriginPeriods,
  DevelopmentPeriod = months(1))
}
\description{
This will create a set of development lags from evaluation dates
}

