\name{CreatePriors}
\alias{CreatePriors}
\title{Create priors}
\usage{
  CreatePriors(dfTriangleData, measureCols, Groups)
}
\arguments{
  \item{OriginPeriod}{Vector of intervals}

  \item{Group}{A vector of groups}

  \item{Variable}{A list of}
}
\value{
  Adjusted triangle
}
\description{
  Create priors
}
\seealso{
  \code{\link{CreateIncremental}},
  \code{\link{CreateCumulative}}
}

