% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstats_plotting.R
\name{draw.Condition}
\alias{draw.Condition}
\title{Draw condition plots}
\usage{
draw.Condition(df, protein)
}
\arguments{
\item{df}{A dataframe with processed/normalized MSstats results}

\item{protein}{A protein identifier}
}
\value{
a ggplot object
}
\description{
Draw an MSstats condition plot for a specific protein. Will perform boxplot for n >= 5, or else a dotplot.
}
\examples{
draw.Condition(rd.proposed$ProcessedData, "S100A12-P80511")

}
