% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstats_plotting.R
\name{draw.Volcano}
\alias{draw.Volcano}
\title{Draw a Volcano Plot}
\usage{
draw.Volcano(df, comparison, sig.p = 0.05)
}
\arguments{
\item{df}{A dataframe containing MSstats comparison results}

\item{comparison}{The comparison to plot}

\item{sig.p}{The adjusted p-value for significance. Default is 0.05}
}
\value{
A ggplot object
}
\description{
Draw a Volcano plot from an MSstats comparison
}
\examples{
draw.Volcano(pd.comparisons$ComparisonResult, 'A3-A1')

}
