% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a mark-recapture model}
\usage{
fit_model(data, ...)
}
\arguments{
\item{data}{A list of data generated by the function
\code{mrmr::clean_data()}.}

\item{...}{Additional parameters to pass to \code{rstan::sampling()}}
}
\value{
A list containing the stanfit object, extracted posterior draws, and
input data.
}
\description{
Fit a mark-recapture model
}
\examples{
\dontrun{
captures <- system.file("extdata", "capture-example.csv", package = "mrmr")
translocations <- system.file("extdata", "translocation-example.csv",
                              package = "mrmr")
surveys <- system.file("extdata", "survey-example.csv", package = "mrmr")
out <- clean_data(captures, translocations, surveys)
fit_model(out, chains = 1, iter = 10)
}
}
