% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SerialCorrelation.R
\name{FitSerialCorrelation}
\alias{FitSerialCorrelation}
\title{Fit the serial correlation in a triangle}
\usage{
FitSerialCorrelation(objTriangleModel)
}
\arguments{
\item{objTriangleModel}{A Triangle model}
}
\description{
Fit the serial correlation in a triangle
}

