\name{FormMeasureNames}
\alias{FormMeasureNames}
\title{Form measures}
\usage{
  FormMeasureNames(Measures, Cumulative = TRUE)
}
\arguments{
  \item{Measures}{A character vector of stochastic measure
  names}

  \item{Cumulative}{Boolean indicating whether the measure
  names are cumulative or incremental}
}
\description{
  Form measures
}

