% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Friedland}
\alias{Friedland}
\title{Friedland data}
\description{
This is a single triangle taken from blah by Jacqueline Friedland, page 65.
}
\references{
\url{http://www.casact.org/library/studynotes/Friedland_estimating.pdf}
}

