% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstats_processing.R
\name{gen.CompMatrix}
\alias{gen.CompMatrix}
\title{Generate a comparison matrix}
\usage{
gen.CompMatrix(mat, lev)
}
\arguments{
\item{mat}{A matrix with desired condition comparisons, 2 or 4 columns}

\item{lev}{Condition levels from the processed data}
}
\value{
An MSstats comparison matrix
}
\description{
Given a matrix of desired comparisons and original group levels, generate the comparison matrix for MSstats
}
\examples{
# Set up a matrix with desired comparisons
# For a double difference, the conditions in positions 1 & 4 will be set to 1
# and conditions in positions 2 & 3 will be set to -1, so these are like
# (Trt.Trted - Trt.Baseline) - (Veh.Trted -  Veh.Baseline)

el <- matrix(c('B.3','B.1','A.3','A.1',
               'B.4','B.1','A.4','A.1',
               'C.3','C.1','A.3','A.1',
               'C.4','C.1','A.4','A.1',
               'C.3','C.1','B.3','B.1',
               'C.4','C.1','B.4','B.1',
               'D.3','D.1','E.3','E.1',
               'D.4','D.1','E.4','E.1',
               'Y.3','Y.1','Z.3','Z.1',
               'Y.4','Y.1','Z.4','Z.1'),
             ncol=4, byrow = T)

orig.levels <- levels(rd.proposed$ProcessedData$GROUP_ORIGINAL)

# Generate MSstats comparison matrix from the graph and original groups
comparisons <- gen.CompMatrix(el, orig.levels)

}
