% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{get_numeric_indices}
\alias{get_numeric_indices}
\title{Get numeric indices from a string representation}
\usage{
get_numeric_indices(string)
}
\arguments{
\item{string}{input string}
}
\description{
Get numeric indices from a string representation
}
