\name{GetStochasticColumnNames}
\alias{GetStochasticColumnNames}
\title{GetStochasticColumnNames}
\usage{
GetStochasticColumnNames(MeasureNames)
}
\arguments{
  \item{MeasureNames}{A character vector of base measure
  names}
}
\value{
A character vector of measure names augmented with the
words Incremental, Cumulative and Prior
}
\description{
GetStochasticColumnNames
}

