% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getY.R
\name{getY}
\alias{getY}
\title{Create an observation matrix}
\usage{
getY(mr, M)
}
\arguments{
\item{mr}{The mark recapture data.}

\item{M}{The number of all zero pseudoindividual detection histories to add.}
}
\value{
A matrix with a row for each individual (+ pseudo-individuals) and a
column for each secondary period, where the entries are binary, indicating
whether an individual was observed during a secondary period.
}
\description{
This function makes an observation matrix Y from a mark recapture dataset.
}

