% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZ.R
\name{getZ}
\alias{getZ}
\title{Create a partly observed occurrence matrix}
\usage{
getZ(Y)
}
\arguments{
\item{Y}{The observation matrix, of class obs_mat.}
}
\value{
A matrix with a row for each individual (+ pseudo-individuals) and a
column for each primary period, where the entries are binary, indicating
whether an individual was present during a primary period. Unknown states
are represented by NA entries.
}
\description{
This function makes an occurrence matrix Z from an observation matrix.
}
