% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OriginPeriod.R
\name{is.OriginPeriod}
\alias{is.OriginPeriod}
\title{is.OriginPeriod}
\usage{
is.OriginPeriod(object)
}
\arguments{
\item{object}{Object to be tested}
}
\value{
TRUE/FALSE
}
\description{
Checks whether the object is an object of the OriginPeriod class.
}

