% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StaticMeasure.R
\name{is.StaticMeasure}
\alias{is.StaticMeasure}
\title{is.StaticMeasure}
\usage{
is.StaticMeasure(object)
}
\arguments{
\item{object}{Object to be}
}
\value{
Boolean
}
\description{
Is the object a StaticMeasure?
}

