% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Triangle.R
\name{is.Triangle}
\alias{is.Triangle}
\title{is.Triangle}
\usage{
is.Triangle(object)
}
\arguments{
\item{object}{The object to be tested}
}
\value{
TRUE if the object is a triangle, FALSE if it is not
}
\description{
Tests whether the object is a triangle
}

