% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatestDiagonal.R
\name{LatestDiagonal}
\alias{LatestDiagonal}
\title{LatestDiagonal}
\usage{
LatestDiagonal(x)
}
\arguments{
\item{x}{a data frame or a triangle}
}
\value{
A data frame
}
\description{
This function will return all of the values for the most recent evaluation date. Note 
that this applies for each origin period individually. For example, if some origin periods have an 
evaluation at December 31, 2010, but others only have evaluations at December 31, 2009, the data frame 
which is returned will have two different evaluation dates present.
}

