% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StaticMeasure.R
\docType{methods}
\name{LevelNames}
\alias{LevelNames}
\title{LevelNames}
\usage{
\S4method{LevelNames}{StaticMeasure}(x)
}
\arguments{
\item{x}{A StaticMeasure object}
}
\description{
LevelNames
}

