% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mr.R
\name{load_mr}
\alias{load_mr}
\title{Load a mark-recapture data set}
\usage{
load_mr(path)
}
\arguments{
\item{path}{The MS Excel file containing data.}
}
\value{
A data.frame
}
\description{
This function loads a mark recapture data set from an Excel file.
}
