% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StaticMeasure.R
\docType{methods}
\name{MeasureNames}
\alias{MeasureNames}
\title{MeasureNames}
\usage{
\S4method{MeasureNames}{StaticMeasure}(x)
}
\arguments{
\item{x}{A StaticMeasure object}
}
\description{
MeasureNames
}

