\name{ModifyTail}
\alias{ModifyTail}
\title{This will eventuall house two functions.
The first has a signature of data frame (or some such) and max tail value.
The second has a signature of TriangleModel and data frame}
\usage{
  ModifyTail(objTriangleModel)
}
\description{
  This will eventuall house two functions. The first has a
  signature of data frame (or some such) and max tail
  value. The second has a signature of TriangleModel and
  data frame
}

