% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmr-package.R
\docType{package}
\name{mrmr-package}
\alias{mrmr-package}
\alias{mrmr}
\title{The 'mrmr' package.}
\description{
CMR models in R.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
