% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Multiline}
\alias{Multiline}
\title{Multiline data}
\description{
This data set is taken from the NAIC data prepared by Glenn Meyers and Peng Shi
}
\references{
\url{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}
}

