\docType{data}
\name{NAIC}
\alias{NAIC}
\title{NAIC workers comp data}
\description{
  This data set contains workers comp development
  information obtained from public filings. The data was
  gathered and prepared by Glenn Meyers and Peng Shi.
}
\references{
  \url{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}
}
\keyword{data}
\keyword{datasets}

