% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOriginPeriods.R
\name{OriginPeriodFromYears}
\alias{OriginPeriodFromYears}
\title{OriginPeriodFromYears}
\usage{
OriginPeriodFromYears(OriginStartYears, OriginLength = years(1),
  StartDay = 1, StartMonth = 1, Verbose = FALSE)
}
\description{
This will construct a set of origin periods from years
}

