\name{PivotTriangle}
\alias{PivotTriangle}
\title{Pivot a triangle}
\usage{
  PivotTriangle(objTriangle, whichMeasure, whichRows,
    whichColumns)
}
\arguments{
  \item{ColumnCategory}{}

  \item{RowCategory}{}

  \item{Value}{}

  \item{data}{}
}
\value{
  Dataframe of pivoted data
}
\description{
  Triangle evaluation dates are established by adding
  development lags to the starting point of the origin
  periods.
}

