% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StaticMeasure.R
\docType{methods}
\name{plot.StaticMeasure}
\alias{plot.StaticMeasure}
\title{StaticMeasure plot}
\usage{
\S4method{plot}{StaticMeasure,missing}(x, Measure, GroupParameter = "Measure",
  FacetFormula)
}
\description{
StaticMeasure plot
}
\details{
A StaticMeasure will always use the OriginPeriod for the x axis.

The GroupParameter controls which column will be used to group the results. Everything else will be used to facet
the plots.
}

