% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotModelFactors.R
\name{PlotModelFactors}
\alias{PlotModelFactors}
\title{PlotModelFactors}
\usage{
PlotModelFactors(objTriangleModel)
}
\arguments{
\item{objTriangleModel}{A TriangleModel object}
}
\value{
A ggplot2 plot object
}
\description{
This function will plot the model factors associated with a triangle model.
}
\seealso{
\code{\link{PlotModelFactors}}
}

