\name{PlotResiduals}
\alias{PlotResiduals}
\title{PlotResiduals}
\usage{
  PlotResiduals(objTriangleModel)
}
\arguments{
  \item{objTriangleModel}{A TriangleModel object}
}
\value{
  This function does not return a value.
}
\description{
  This will produce a 2x2 set of residual graphs.
}
\details{
  This function will produce four charts.
}

