% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SerialCorrelation.R
\name{plotSerialCorrelation}
\alias{plotSerialCorrelation}
\title{Plot the serial correlation in a triangle}
\usage{
plotSerialCorrelation(objTriangleModel)
}
\arguments{
\item{objTriangleModel}{A Triangle model}
}
\description{
Plot the serial correlation in a triangle
}

