\name{plotTriangle}
\alias{plotTriangle}
\title{plot.Triangle}
\usage{
  plotTriangle(x, y, Predictor, Response,
    Group = "OriginPeriodStart", Lines = TRUE,
    FitLines = FALSE, ...)
}
\arguments{
  \item{x}{Triangle y The measure being plotted Predictor
  The variable used to predict the response Group The name
  of the group column used to group the data. The default
  is OriginPeriodStart Lines Draw lines to connect the
  observations? FitLines Draw a line of best fit? Note that
  fit lines will have an intercept}
}
\description{
  plot.Triangle
}

