\name{plotTriangleModel}
\alias{plotTriangleModel}
\title{plotTriangleModel}
\usage{
  plotTriangleModel(x, y, ...)
}
\arguments{
  \item{OriginPeriod}{Vector of intervals}
}
\value{
  A vector of intervals
}
\description{
  plotTriangleModel
}
\seealso{
  \code{\link{CreateCumulative}},
  \code{\link{CreatePriors}}
}

