% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstats_processing.R
\name{preprocess.MSstats}
\alias{preprocess.MSstats}
\title{Preprocess Skyline files for MSstats analysis}
\usage{
preprocess.MSstats(input.csv, annot.df)
}
\arguments{
\item{input.csv}{csv file created by MSstats Skyline report}

\item{annot.df}{data frame with 'Condition', 'Run', and 'BioReplicate' columns}
}
\value{
dataframe suitable for MSstats processing
}
\description{
Combines feature intensities and annotation to yield appropriate MSstats input
}
\examples{
rd.msstats <- preprocess.MSstats(paste0(proj.dir,'Data/MSstats_Input.csv'), annot)

}
