% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstats_plotting.R
\name{print.Volcanos}
\alias{print.Volcanos}
\title{Print all Volcano Plots}
\usage{
\method{print}{Volcanos}(df)
}
\arguments{
\item{df}{A dataframe with MSstats comparison results}
}
\description{
Prints Volcano plots for all performed comparisons to the screen
}
\examples{
print.Volcanos(pd.comparisons$ComparisonResult)

}
