\name{ProjectToDate}
\alias{ProjectToDate}
\title{ProjectToDate}
\usage{
ProjectToDate(objTriangleModel, lOriginYears, AsOfDate)
}
\arguments{
  \item{objTriangleModel}{A TriangleModel object}

  \item{lOriginYears}{A list of origin years}

  \item{AsOfDate}{A date to which to project}
}
\value{
A data frame which has projected dates and columns for the
new stochastic values
}
\description{
This function
}

