% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectToDev.R
\name{ProjectToDev}
\alias{ProjectToDev}
\title{Project to a specific development lag}
\usage{
ProjectToDev(objTriangleModel, lOriginYears, MaxDev)
}
\description{
Given a triangle model, this will project results to a specific development lag
}

