\name{RenameCoefficients}
\alias{RenameCoefficients}
\title{RenameCoefficients}
\usage{
RenameCoefficients(Fit, Names)
}
\arguments{
  \item{x}{a data frame or a triangle}
}
\value{
A data frame
}
\description{
This function will return all of the values for the most
recent evaluation date. Note that this applies for each
origin period individually. For example, if some origin
periods have an evaluation at December 31, 2010, but others
only have evaluations at December 31, 2009, the data frame
which is returned will have two different evaluation dates
present.
}

