% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{StaticMeasure-class}
\alias{StaticMeasure-class}
\title{StaticMeasure class}
\description{
A StaticMeasure holds multilevel/multivariate data associated with an OriginPeriod.
}
\details{
A StaticMeasure is effectively a time series, or set of time series. Because each value of a StaticMeasure
corresponds to an element of an OriginPeriod- which represents an interval of time- the StaticMeasure does not represent
an observation at a single point in time. In an insurance context, common StaticMeasures would include written or earned
premium, written or earned exposure, number of insured locations for a fixed duration and the like. In-force premium or
in-force exposure would not constitute a StaticMeasure as this is a measure at a single point in time.

A StaticMeasure is so called to distinguish it from a StochasticMeasure. A StochasticMeasure also has a correspondence
to a single element of an OriginPeriod, but its value changes from one observation to the next. This would be the case for
number of claims, paid losses and the like. This would also be true for loss-sensitive premium such as exists in a retrospectively
rated policy.

A StaticMeasure object may hold more than one measure for each time period. Earned premium and policy count may both be stored
in a single object, for example. This is done to facilitate a multivariate approach to the loss reserving problem. Stochastic
responses may depend on more than one variable.

\strong{Levels}
Each StaticMeasure may be assigned level parameters which enable multiple sets of measures to be stored in the
same object. This will, in turn, facilitate construction of a hierarchical analysis model. The levels may be
of any arbitraty complexity, irrespective of the number of elements of the StaticMeasure object. For example, one
may have a single StaticMeasure which uses more than one level element: GL, prem/ops, California, for example.
The object is constrained in that there must be enough levels to describe each element of the object. The constructor examples
should make this clear.

If the user does not specify level attributes, the default value of "All" is used.

The number of levels of a StaticMeasure will determine its length. A StaticMeasure may be thought of as containing a list, wherein each
element of the list holds a data frame. The OriginPeriod and Measure properties ensure that each data frame will contain comparable
information.

\strong{StaticMeasure construction}
To construct

Factors will be converted to characters by default. This is advisable.

\strong{Member access}
Elements of a StaticMeasure object are accessed in a manner similar to those for data frames. However, a StaticMeasure has two, not
three dimensions. One way of thinking of this is that a StaticMeasure is a list of data frames. The indexing convention is as follows:
Because the "data frame" may be thought of as having rows specific by OriginPeriod and columns specified by the names of measured
values, the i and i indices are used to specify these two dimensions. The third dimension is used to identify the Level. For a
StochasticMeasure object, a fourth dimension is used to identify the evalution date, or development lag.
}
\seealso{
\code{\link{StaticMeasureConstructor}}
}

