% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StaticMeasure.R
\docType{methods}
\name{StaticMeasure}
\alias{StaticMeasure}
\title{StaticMeasure}
\usage{
\S4method{StaticMeasure}{OriginPeriod}(OriginPeriod, Measure, Level, Data,
  OriginPeriodSort)
}
\arguments{
\item{OriginPeriod}{OriginPeriod object}

\item{Measure}{Vector of names of columns}

\item{Level}{List of levels}

\item{Data}{A data frame}
}
\description{
StaticMeasure
}

