% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StochasticMeasure.R
\docType{class}
\name{StochasticMeasure-class}
\alias{StochasticMeasure-class}
\title{StochasticMeasure class}
\description{
A StochasticMeasure holds multilevel/multivariate data associated with an OriginPeriod.
}
\details{
A StochasticMeasure resembles longitudinal data.

A StochasticMeasure is so called to distinguish it from a StaticMeasure. A StochasticMeasure has a correspondence
to a single element of an OriginPeriod, but its value changes from one observation to the next. This would be the case for
number of claims, paid losses and the like. This would also be true for loss-sensitive premium such as exists in a retrospectively
rated policy.


\strong{Levels}
Each StochasticMeasure may be assigned level parameters which enable multiple sets of measures to be stored in the
same object. This will, in turn, facilitate construction of a hierarchical analysis model. The levels may be
of any arbitraty complexity, irrespective of the number of elements of the StochasticMeasure object. For example, one
may have a single StochasticMeasure which uses more than one level element: GL, prem/ops, California, for example.
The object is constrained in that there must be enough levels to describe each element of the object. The constructor examples
should make this clear.

If the user does not specify level attributes, the default value of "All" is used.

\strong{StochasticMeasure construction}
To construct

\strong{Member access}
Elements of a StochasticMeasure object are accessed in a manner similar to those for data frames. However, a StochasticMeasure has two, not
three dimensions. One way of thinking of this is that a StochasticMeasure is a list of data frames. The indexing convention is as follows:
Because the "data frame" may be thought of as having rows specific by OriginPeriod and columns specified by the names of measured
values, the i and i indices are used to specify these two dimensions. The third dimension is used to identify the Level. For a
StochasticMeasure object, a fourth dimension is used to identify the evalution date, or development lag.
}

