\name{summaryTriangleModel}
\alias{summaryTriangleModel}
\title{summaryTriangleModel}
\usage{
  summaryTriangleModel(object, ...)
}
\arguments{
  \item{OriginPeriod}{Vector of intervals}
}
\value{
  A vector of intervals
}
\description{
  summaryTriangleModel
}
\seealso{
  \code{\link{CreateCumulative}},
  \code{\link{CreatePriors}}
}

