% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_table.R
\name{survival_table}
\alias{survival_table}
\title{Generate a table of introduced survival over time}
\usage{
survival_table(model)
}
\arguments{
\item{model}{A model objectgenerated by \code{mrmr:fit_model()}.}
}
\value{
A data.frame with columns year_since_introduction,
lo_survival, med_survival, and hi_survival,
where lo, med, and hi represent the 2.5%, 50%, and 97.5% posterior quantiles
for survival in the years follwing introduction.
}
\description{
This function generates a table of survival estimates for introduced
individuals after introduction. These estimates average over individuals,
and over primary periods within years, so that the values are comparable
among sites with one or multiple introduced cohorts, and
a variable number of primary periods per year.
}
\examples{
\dontrun{
captures <- system.file("extdata", "capture-example.csv", package = "mrmr")
translocations <- system.file("extdata", "translocation-example.csv",
                              package = "mrmr")
surveys <- system.file("extdata", "survey-example.csv", package = "mrmr")
out <- clean_data(captures, translocations, surveys)
model <- fit_model(out, chains = 1, iter = 10)
survival_table(model)
}
}
