% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Triangle-class}
\alias{Triangle-class}
\title{Triangle class}
\description{
Triangle is an S4 class used to store aggregated loss data. It is composed of a StochasticMeasure and
a StaticMeasure object.
}
\details{
\strong{Triangle construction}
}
\seealso{
\code{\link{Triangle}}
}

