% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleModel.R
\docType{class}
\name{TriangleModel-class}
\alias{TriangleModel-class}
\title{TriangleModel class}
\description{
Triangle is an S4 class used to store a model fit to a Triangle object.
}
\details{
Some stuff
}
\seealso{
\code{\link{Triangle-class}}
}

