% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleProjection.R
\docType{class}
\name{TriangleProjection-class}
\alias{TriangleProjection-class}
\title{TriangleProjection class}
\description{
TriangleProjection is an S4 class used to project values.
}
\seealso{
\code{\link{newTriangle}}
}

