% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Load a mark-recapture data set}
\usage{
clean_data(captures, translocations, surveys)
}
\arguments{
\item{captures}{Character string path to a capture-recapture detection
data csv file.}

\item{translocations}{Character string path to translocation data csv file.}

\item{surveys}{Character string path to survey data csv file.}
}
\value{
A list containing the data frames resulting from the capture,
translocation, and survey data, along with a list of data formatted for
use in a mark recapture model (with name "stan_d").
}
\description{
This function loads and parses a mark recapture data set. It assumes that
three files are available, specifying capture, translocation, and survey
data.
}
\examples{
captures <- system.file("extdata", "capture-example.csv", package = "mrmr")
translocations <- system.file("extdata", "translocation-example.csv",
                              package = "mrmr")
surveys <- system.file("extdata", "survey-example.csv", package = "mrmr")
clean_data(captures, translocations, surveys)
}
