% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteTriangle.R
\name{CompleteTriangle}
\alias{CompleteTriangle}
\title{CompleteTriangle}
\usage{
CompleteTriangle(objProjection)
}
\arguments{
\item{objProjection}{A TriangleProjection object}
}
\value{
A data frame with the sample data (the "upper triangle") bound with the projected data (the "lower triangle").
}
\description{
This function will bind the projected values to the base triangle data for a "complete" triangle. This facilitates comparison of ultimates between multiple TriangleModels.
}

