% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleAdjustMeasures.R
\name{CreateCumulative}
\alias{CreateCumulative}
\title{Create cumulative}
\usage{
CreateCumulative(dfTriangleData, measureCols, Groups)
}
\arguments{
\item{dfTriangleData}{A data frame of triangle variables}

\item{measureCols}{A character vector which holds column names identifying stochastic measures}

\item{Groups}{A character vector which holds column names identifying groups}
}
\value{
A data frame of measures with cumulatives included
}
\description{
Create cumulative
}
\seealso{
\code{\link{CreateIncrementals}}, \code{\link{CreatePriors}}
}

