% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleAdjustMeasures.R
\name{CreateIncrementals}
\alias{CreateIncrementals}
\title{Create incrementals}
\usage{
CreateIncrementals(dfTriangleData, measureCols, Groups)
}
\arguments{
\item{dfTriangleData}{A data frame of triangle variables}

\item{measureCols}{A character vector which holds column names identifying stochastic measures}

\item{Groups}{A character vector which holds column names identifying groups}
}
\value{
A data frame of measures which includes incrementals
}
\description{
Create incrementals
}
\seealso{
\code{\link{CreateCumulative}}, \code{\link{CreatePriors}}
}

