% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDevelopmentLags.R
\name{CreateLagsFromIntegers}
\alias{CreateLagsFromIntegers}
\title{CreateLagsFromIntegers}
\usage{
CreateLagsFromIntegers(LagValues, DevelopmentPeriod = months(1),
  Verbose = TRUE)
}
\description{
This will create a set of development lags from integers
}

