% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleAdjustMeasures.R
\name{FormMeasureNames}
\alias{FormMeasureNames}
\title{Form measures}
\usage{
FormMeasureNames(Measures, Cumulative = TRUE)
}
\arguments{
\item{Measures}{A character vector of stochastic measure names}

\item{Cumulative}{Boolean indicating whether the measure names are cumulative or incremental}
}
\description{
Form measures
}

