% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleAdjustMeasures.R
\name{GetStochasticColumnNames}
\alias{GetStochasticColumnNames}
\title{GetStochasticColumnNames}
\usage{
GetStochasticColumnNames(MeasureNames)
}
\arguments{
\item{MeasureNames}{A character vector of base measure names}
}
\value{
A character vector of measure names augmented with the words Incremental, Cumulative and Prior
}
\description{
GetStochasticColumnNames
}

