% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.OriginPeriod}
\alias{is.OriginPeriod}
\title{is.OriginPeriod}
\usage{
is.OriginPeriod(object)
}
\arguments{
\item{object}{Object to be tested}
}
\value{
TRUE/FALSE
}
\description{
Checks whether the object is an object of the OriginPeriod class.
}

