% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOriginPeriods.R
\name{OriginPeriodFromDates}
\alias{OriginPeriodFromDates}
\title{OriginPeriodFromDates}
\usage{
OriginPeriodFromDates(OriginStart, OriginEnd, Verbose = FALSE)
}
\description{
This will extract a set of origin period values from a range of dates
}

