% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot.StaticMeasure}
\alias{plot.StaticMeasure}
\title{StaticMeasure plot}
\usage{
\S4method{plot}{StaticMeasure,missing}(x, Measure, GroupParameter = "Measure",
  FacetFormula)
}
\arguments{
\item{x}{the coordinates of points in the plot. Alternatively, a
      single plotting structure, function or \emph{any \R object with a
        \code{plot} method} can be provided.}
}
\description{
StaticMeasure plot
}
\details{
A StaticMeasure will always use the OriginPeriod for the x axis.

The GroupParameter controls which column will be used to group the results. Everything else will be used to facet
the plots.
}

