% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotModelGoF.R
\name{PlotModelGoF}
\alias{PlotModelGoF}
\title{PlotModelGoF}
\usage{
PlotModelGoF(objTriangleModel)
}
\arguments{
\item{objTriangleModel}{A TriangleModel object}
}
\value{
A vector of intervals
}
\description{
This function will plot the F distribution associated with the TriangleModel, along with a vertical line indicating the 
F statistic for this model.
}
\seealso{
\code{\link{PlotModelFactors}}
}

