% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotResiduals.R
\name{PlotResiduals}
\alias{PlotResiduals}
\title{PlotResiduals}
\usage{
PlotResiduals(objTriangleModel)
}
\arguments{
\item{objTriangleModel}{A TriangleModel object}
}
\value{
This function does not return a value.
}
\description{
This will produce a 2x2 set of residual graphs.
}
\details{
This function will produce four charts.
}

