\name{plotTriangleModel}
\alias{plotTriangleModel}
\title{plotTriangleModel}
\usage{
  plotTriangleModel(x, y, ...)
}
\arguments{
  \item{x}{A TriangleModel object y Nuisance parameter}
}
\value{
  None
}
\description{
  plotTriangleModel
}
\seealso{
  \code{\link{PlotModelGoF}},
  \code{\link{PlotModelFactors}}
}

