\name{plotTriangleModel}
\alias{plotTriangleModel}
\title{plotTriangleModel}
\usage{
plotTriangleModel(objTriangleModel)
}
\arguments{
  \item{objTriangleModel}{A TriangleModel object}
}
\value{
None
}
\description{
plotTriangleModel
}
\seealso{
\code{\link{PlotModelGoF}}, \code{\link{PlotModelFactors}}
}

