% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTriangleModel.R
\name{plotTriangleModel}
\alias{plotTriangleModel}
\title{plotTriangleModel}
\usage{
plotTriangleModel(objTriangleModel)
}
\arguments{
\item{objTriangleModel}{A TriangleModel object}
}
\value{
None
}
\description{
plotTriangleModel
}
\seealso{
\code{\link{PlotModelGoF}}, \code{\link{PlotModelFactors}}
}

