\name{summaryTriangleModel}
\alias{summaryTriangleModel}
\title{summaryTriangleModel}
\usage{
  summaryTriangleModel(x, ...)
}
\arguments{
  \item{x}{TriangleModel object}
}
\value{
  A vector of intervals
}
\description{
  summaryTriangleModel
}
\seealso{
  \code{\link{CreateCumulative}},
  \code{\link{CreatePriors}}
}

