% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTriangleModel.R
\name{summaryTriangleModel}
\alias{summaryTriangleModel}
\title{summaryTriangleModel}
\usage{
summaryTriangleModel(objTriangleModel)
}
\arguments{
\item{objTriangleModel}{TriangleModel object}
}
\value{
A vector of intervals
}
\description{
summaryTriangleModel
}
\seealso{
\code{\link{CreateCumulative}}, \code{\link{CreatePriors}}
}

