% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Triangle.R
\docType{class}
\name{Triangle-class}
\alias{Triangle-class}
\title{Triangle class}
\description{
Triangle is an S4 class used to store aggregated loss data. It is composed of a StochasticMeasure and
a StaticMeasure object.
}
\details{
\strong{Triangle construction}
}

