\name{Triangle}
\alias{Triangle}
\title{Create a Triangle object
Triangle creates a new Triangle}
\usage{
  Triangle(TriangleData, TriangleName, OriginPeriodColumn,
    OriginPeriodType = "accident",
    OriginPeriodInterval = years(1), DevelopmentColumn,
    DevelopmentInterval = years(1), MeasureMeta)
}
\arguments{
  \item{TriangleData}{A dataframe}

  \item{TriangleName}{The name of the triangle}

  \item{OriginPeriodType}{A character string which
  describes the type of origin period.}

  \item{OriginPeriodInterval}{}
}
\description{
  Create a Triangle object Triangle creates a new Triangle
}

